Make mDNS part of DiscoveryBehaviour
